Separate client-manager and manager-workers communications
