+++ task_2_1_2/src/main/java/kozlov/kirill/sockets/exceptions/InternalWorkerErrorException.java
@@ -1,0 +1,3 @@
+package kozlov.kirill.sockets.exceptions;
+
+public class InternalWorkerErrorException extends Exception {}
